+++ Package/Tests/UnityTests/AwaitableExtensionsTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4bde3a1b457c974db3058515f176fd5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
