+++ Package/UnityHelpers/2018.3.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e246d8de5a314b243986cafeb38dce8b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
