+++ Package/UnityHelpers/2018.3/Internal.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d7848a3b392d99d4687c5c1d5d74b2f6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
