+++ Package/UnityHelpers/2018.3/Internal/PromiseYielderInternal.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4955723c975cace4dabfa68b6c2031de
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
