+++ Package/UnityHelpers/2018.3/PromiseYieldInstruction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f3c04361e58248243b1503ab64ae4be8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
