+++ Package/UnityHelpers/2018.3/PromiseYielder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a68976b1ce678494fac1c860828ed5a6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
