+++ Package/UnityHelpers/2018.3/ProtoPromise.UnityHelpers.2018.3.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5aa80380c7d32394bb76a6ad338b52bc
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
