+++ Package/UnityHelpers/2023.1.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a60cd0d1eecbea5459581f4becf146a9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
