+++ Package/UnityHelpers/2023.1/ProtoPromise.UnityHelpers.2023.1.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "ProtoPromise.UnityHelpers.2023.1",
+    "rootNamespace": "",
+    "references": [
+        "ProtoPromise"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
