+++ Package/UnityHelpers/2023.1/ProtoPromise.UnityHelpers.2023.1.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d3d82f46d058d824db399a37a1776659
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
