--- Package/UnityHelpers/Internal/PromiseYielderInternal.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4955723c975cace4dabfa68b6c2031de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
