--- Package/UnityHelpers/PromiseYieldExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac86b745a86e1df48898f12bf0d80baa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
