--- Package/UnityHelpers/ProtoPromiseUnityHelpers.asmdef
@@ -1,14 +1,0 @@
-{
-    "name": "ProtoPromiseUnityHelpers",
-    "references": [
-        "ProtoPromise"
-    ],
-    "optionalUnityReferences": [],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": true,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": []
-}
