--- Package/UnityHelpers/ProtoPromiseUnityHelpers.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5aa80380c7d32394bb76a6ad338b52bc
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
