+++ ProtoPromise.UnityHelpers/2023.1/nuget/ProtoPromise.UnityHelpers.2023.1.nuspec
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<package >
+  <metadata minClientVersion="2.5">
+    <id>ProtoPromise.UnityHelpers.2023.1</id>
+    <version>{VERSION}</version>
+    <title>ProtoPromise Unity Helpers</title>
+    <authors>Tim Cassell</authors>
+    <owners>Tim Cassell</owners>
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <license type="expression">MIT</license>
+    <projectUrl>https://github.com/timcassell/ProtoPromise</projectUrl>
+    <repository type="git" url="https://github.com/timcassell/ProtoPromise" />
+    <description>ProtoPromise helpers for UnityEngine.</description>
+    <releaseNotes></releaseNotes>
+    <copyright>Copyright ©2024 Tim Cassell</copyright>
+    <tags>promise promises task tasks csharp unity dotnet mono coroutine coroutines concurrency concurrent parallel asynchronous async await thread threads threading then thenable callback callbacks</tags>
+
+    <dependencies>
+      <group targetFramework="netstandard2.1">
+        <dependency id="ProtoPromise" version="{VERSION}" exclude="Build,Analyzers" />
+        <dependency id="ProtoPromise.UnityHelpers.2018.3" version="{VERSION}" exclude="Build,Analyzers" />
+      </group>
+    </dependencies>
+
+    <readme>readme.md</readme>
+  </metadata>
+
+  <files>
+    <file src="..\bin\Release\netstandard2.1\**" target="lib/netstandard2.1/Release" />
+    
+    <file src="readme.md" target="" />
+  </files>
+</package>
