+++ ProtoPromise.UnityHelpers/2023.1/nuget/readme.md
@@ -1,0 +1,3 @@
+Helpers to assist with using Promises in UnityEngine 2023.1 or newer.
+
+Note: it is not recommended to use this nuget package. Instead, you should use the Unity package available on [GitHub](https://github.com/timcassell/ProtoPromise/releases) or the [Unity Asset Store](https://assetstore.unity.com/packages/tools/integration/protopromise-181997).
