+++ ProtoPromise.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProtoPromiseTests", "ProtoP
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProtoPromise", "ProtoPromise\ProtoPromise.csproj", "{2A0BBD14-FA01-467C-8A42-55C5C287F66D}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProtoPromiseUnityHelpers", "ProtoPromiseUnityHelpers\ProtoPromiseUnityHelpers.csproj", "{5B1E6550-0342-4240-AF91-F8059F84BF67}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{9790CDFB-EFE9-4657-9F36-577112C30BE5}"
 	ProjectSection(SolutionItems) = preProject
 		Package\CHANGELOG.md = Package\CHANGELOG.md
@@ -20,6 +18,10 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProtoPromise.Analyzer", "ProtoPromise.Analyzer\ProtoPromise.Analyzer\ProtoPromise.Analyzer.csproj", "{B932E015-4D00-49FE-AA00-CEA4FB9EAEF9}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProtoPromise.Analyzer.Test", "ProtoPromise.Analyzer\ProtoPromise.Analyzer.Test\ProtoPromise.Analyzer.Test.csproj", "{F978509B-1313-4EBE-88EF-D88DE4E30DFA}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProtoPromise.UnityHelpers.2018.3", "ProtoPromise.UnityHelpers\2018.3\ProtoPromise.UnityHelpers.2018.3.csproj", "{3F910193-841A-4DC0-AF4F-8B34F59ACD25}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ProtoPromise.UnityHelpers.2023.1", "ProtoPromise.UnityHelpers\2023.1\ProtoPromise.UnityHelpers.2023.1.csproj", "{A63A6C25-72C8-41D0-9077-10F49E59B578}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -35,10 +37,6 @@ Global
 		{2A0BBD14-FA01-467C-8A42-55C5C287F66D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{2A0BBD14-FA01-467C-8A42-55C5C287F66D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{2A0BBD14-FA01-467C-8A42-55C5C287F66D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5B1E6550-0342-4240-AF91-F8059F84BF67}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5B1E6550-0342-4240-AF91-F8059F84BF67}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5B1E6550-0342-4240-AF91-F8059F84BF67}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5B1E6550-0342-4240-AF91-F8059F84BF67}.Release|Any CPU.Build.0 = Release|Any CPU
 		{B932E015-4D00-49FE-AA00-CEA4FB9EAEF9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{B932E015-4D00-49FE-AA00-CEA4FB9EAEF9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B932E015-4D00-49FE-AA00-CEA4FB9EAEF9}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -47,6 +45,14 @@ Global
 		{F978509B-1313-4EBE-88EF-D88DE4E30DFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F978509B-1313-4EBE-88EF-D88DE4E30DFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{F978509B-1313-4EBE-88EF-D88DE4E30DFA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3F910193-841A-4DC0-AF4F-8B34F59ACD25}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3F910193-841A-4DC0-AF4F-8B34F59ACD25}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3F910193-841A-4DC0-AF4F-8B34F59ACD25}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3F910193-841A-4DC0-AF4F-8B34F59ACD25}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A63A6C25-72C8-41D0-9077-10F49E59B578}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A63A6C25-72C8-41D0-9077-10F49E59B578}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A63A6C25-72C8-41D0-9077-10F49E59B578}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A63A6C25-72C8-41D0-9077-10F49E59B578}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
