--- ProtoPromiseUnityHelpers/ProtoPromiseUnityHelpers.csproj
@@ -1,62 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <!--We target .Net Standard 2.0 to support old Unity versions, and we also target .Net Standard 2.1 to use function pointers.-->
-    <TargetFrameworks>netstandard2.0;netstandard2.1</TargetFrameworks>
-    <Version>2.7.0</Version>
-    <!--In case any IL2CPP issues need to be resolved like in the core library.-->
-    <DefineConstants>$(DefineConstants);ENABLE_IL2CPP</DefineConstants>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <GenerateAssemblyInfo>true</GenerateAssemblyInfo>
-    <GenerateDocumentationFile>true</GenerateDocumentationFile>
-    <ProduceReferenceAssembly>false</ProduceReferenceAssembly>
-  </PropertyGroup>
-
-  <PropertyGroup>
-    <AllowUnsafeBlocks>false</AllowUnsafeBlocks>
-    <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <!-- Remove unnecessary suppression -->
-    <NoWarn>IDE0079</NoWarn>
-  </PropertyGroup>
-
-  <Choose>
-    <When Condition="'$(TargetFramework)'=='netstandard2.0'">
-      <PropertyGroup>
-        <LangVersion>7.3</LangVersion>
-      </PropertyGroup>
-    </When>
-    <Otherwise>
-      <PropertyGroup>
-        <!-- The language version Unity supports with netstandard2.1. -->
-        <LangVersion>9</LangVersion>
-        <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-        <!--We check UNITY_2021_2_OR_NEWER to see if C#9 is supported, since Unity doesn't have a language define constant for it.-->
-        <DefineConstants>$(DefineConstants);UNITY_2021_2_OR_NEWER</DefineConstants>
-      </PropertyGroup>
-    </Otherwise>
-  </Choose>
-
-  <ItemGroup>
-    <Compile Include="..\Package\UnityHelpers\**/*.cs" />
-    <Compile Remove="nuget\**" />
-    <EmbeddedResource Remove="nuget\**" />
-    <None Remove="nuget\**" />
-    <None Remove="UnityEngine.dll" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ProtoPromise\ProtoPromise.csproj">
-      <Private>false</Private>
-    </ProjectReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Reference Include="UnityEngine">
-      <HintPath>UnityEngine.dll</HintPath>
-      <Private>false</Private>
-    </Reference>
-  </ItemGroup>
-
-</Project>
