--- ProtoPromiseUnityHelpers/nuget/ProtoPromiseUnityHelpers.nuspec
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<package >
-  <metadata minClientVersion="2.5">
-    <id>ProtoPromiseUnityHelpers</id>
-    <version>{VERSION}</version>
-    <title>ProtoPromise Unity Helpers</title>
-    <authors>Tim Cassell</authors>
-    <owners>Tim Cassell</owners>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <license type="expression">MIT</license>
-    <projectUrl>https://github.com/timcassell/ProtoPromise</projectUrl>
-    <repository type="git" url="https://github.com/timcassell/ProtoPromise" />
-    <description>ProtoPromise helpers for UnityEngine.</description>
-    <releaseNotes></releaseNotes>
-    <copyright>Copyright ©2022 Tim Cassell</copyright>
-    <tags>promise promises task tasks csharp unity dotnet mono coroutine coroutines concurrency concurrent parallel asynchronous async await thread threads threading then thenable callback callbacks</tags>
-
-    <dependencies>
-      <group targetFramework="net35">
-        <dependency id="ProtoPromise" version="{VERSION}" exclude="Build,Analyzers" />
-      </group>
-      <group targetFramework="netstandard2.1">
-        <dependency id="ProtoPromise" version="{VERSION}" exclude="Build,Analyzers" />
-      </group>
-    </dependencies>
-
-    <readme>readme.md</readme>
-  </metadata>
-
-  <files>
-    <file src="..\bin\Release\net35\**" target="lib/net35/Release" />
-    <file src="..\bin\Release\netstandard2.1\**" target="lib/netstandard2.1/Release" />
-    
-    <file src="readme.md" target="" />
-  </files>
-</package>
