--- ProtoPromiseUnityHelpers/nuget/readme.md
@@ -1,3 +1,0 @@
-Helpers to assist with using Promises in UnityEngine.
-
-Note: it is not recommended to use this nuget package. Instead, you should use the Unity package available on [GitHub](https://github.com/timcassell/ProtoPromise/releases) or the [Unity Asset Store](https://assetstore.unity.com/packages/tools/integration/protopromise-181997).
